+++ cli/prev.c
@@ -4,15 +4,17 @@
 int tman_cli_prev(int argc, char **argv, tman_ctx_t *ctx)
 {
     char c, *errfmt;
-    int showhelp, status;
+    int quiet, showhelp, status;
     struct tman_id_prev_opt opt;
 
-    showhelp = FALSE;
+    quiet = showhelp = FALSE;
     errfmt =  "cannot switch: %s";
-    while ((c = getopt(argc, argv, ":h")) != -1) {
+    while ((c = getopt(argc, argv, ":hq")) != -1) {
         switch (c) {
             case 'h':
                 showhelp = TRUE; break;
+            case 'q':
+                quiet = TRUE; break;
             case ':':
                 return elog(1, "option `-%c' requires an argument", optopt);
             default:
@@ -22,7 +24,9 @@ int tman_cli_prev(int argc, char **argv, tman_ctx_t *ctx)
 
     if (showhelp == TRUE)
         return help_usage("prev");
-    if ((status = tman_id_prev(ctx, &opt)) != TMAN_OK)
-        return elog(status, errfmt, tman_strerror());
-    return tman_pwd();
+    if ((status = tman_id_prev(ctx, &opt)) != TMAN_OK) {
+        if (quiet == FALSE)
+            elog(status, errfmt, tman_strerror());
+    }
+    return status == TMAN_OK ? tman_pwd() : status;
 }