+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `tman`: Add support to not switch to newly created project
 - `cli`: Add extra space between task/project tag and description
 - `lib/dir`: Add support to create project hidden directory
+- `cli/prev`: Add option quiet to suppress error messages
 
 ### Changed